+++ src/main/java/racingcar/domain/dto/FinalResult.java
@@ -1,6 +1,7 @@
-package racingcar.domain;
+package racingcar.domain.dto;
 
 import java.util.List;
+import racingcar.domain.Car;
 
 public class FinalResult {
     private final List<Car> carList;
